Slim ARM kernel arch via bluekernel_arch
